requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-lru==1.0.3
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.1
future==0.18.2
idna==3.4
kafka-python==2.0.2
multidict==6.0.2
oauthlib==3.2.2
pyaml==21.10.1
pyspark==3.1.3
python-twitter==3.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
tweepy==4.12.0
twitter==1.19.6
urllib3==1.26.12
yarl==1.8.1
